requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.3
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
factory-boy==3.2.0
Faker==9.3.1
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pipdeptree==2.2.0
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
regex==2021.10.8
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.1
typing-extensions==3.10.0.2
